apps/backend/pyproject.toml
[project]
name = "smarthome-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "b1ngoinho11",email = "budthbundit.thayaurai@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref (==3.10.0)",
    "django (==5.2.7)",
    "django-debug-toolbar (==6.0.0)",
    "django-cors-headers (>=4.3.0,<5.0.0)",
    "djangorestframework (==3.16.1)",
    "psycopg2-binary (==2.9.11)",
    "sqlparse (==0.5.3)",
    "djangorestframework-gis (>=1.2.0,<2.0.0)",
    "zodb (>=5)",
    "persistent (>=5)",
    "btrees (>=5)",
    "transaction (>=5.0,<6.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "channels (>=4.3.2,<5.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "openai (>=2.16.0,<3.0.0)",
    "google-genai (>=1.61.0,<2.0.0)",
    "groq (>=1.0.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "coverage (>=7.13.5,<8.0.0)"
]
